Play with the layout.
