+++ client/dashboard/sites/performance/core-metrics-chart.tsx
@@ -1,22 +1,32 @@
 import { LineChart } from '@automattic/charts';
+import {
+	__experimentalText as Text,
+	__experimentalHStack as HStack,
+	__experimentalVStack as VStack,
+} from '@wordpress/components';
+import { __, sprintf } from '@wordpress/i18n';
+import { StatusIndicator } from './core-web-vitals/status-indicator';
 
 interface ChartProps {
 	data?: any;
+	activeTab: string;
+	metricsThresholds: any;
 }
 
-export default function CoreMetricsChart( { data }: ChartProps ) {
-	
+export default function CoreMetricsChart( { data, activeTab, metricsThresholds }: ChartProps ) {
+	const { good, needsImprovement, bad } = metricsThresholds[ activeTab ];
+
 	// Generate dummy data for the last 30 days
 	const generateDummyData = ( baseValue: number, variance: number ) => {
 		const now = new Date();
 		return Array.from( { length: 30 }, ( _, index ) => {
 			const date = new Date( now );
 			date.setDate( date.getDate() - ( 29 - index ) );
-			
+
 			// Add some realistic variance to the data
 			const randomVariance = ( Math.random() - 0.5 ) * variance;
 			const value = Math.max( 0, baseValue + randomVariance );
-			
+
 			return {
 				date: date,
 				value: Math.round( value * 100 ) / 100, // Round to 2 decimal places
@@ -24,9 +34,9 @@ export default function CoreMetricsChart( { data }: ChartProps ) {
 		} );
 	};
 
-	const _data: Array<{
+	const _data: Array< {
 		label: string;
-		data: Array<{ date: Date; value: number }>;
+		data: Array< { date: Date; value: number } >;
 		options: {
 			gradient: {
 				from: string;
@@ -39,7 +49,7 @@ export default function CoreMetricsChart( { data }: ChartProps ) {
 				color: string;
 			};
 		};
-	}> = [
+	} > = [
 		{
 			label: 'Largest Contentful Paint (LCP)',
 			data: generateDummyData( 2.5, 0.8 ), // LCP values around 2.5s with variance
@@ -56,48 +66,111 @@ export default function CoreMetricsChart( { data }: ChartProps ) {
 				},
 			},
 		},
-		{
-			label: 'First Input Delay (FID)',
-			data: generateDummyData( 100, 50 ), // FID values around 100ms with variance
-			options: {
-				gradient: {
-					from: '#5BA300',
-					to: '#5BA300',
-					fromOpacity: 0.2,
-					toOpacity: 0,
-				},
-				stroke: '#5BA300',
-				legendShapeStyle: {
-					color: '#5BA300',
-				},
-			},
-		},
-		{
-			label: 'Cumulative Layout Shift (CLS)',
-			data: generateDummyData( 0.1, 0.05 ), // CLS values around 0.1 with variance
-			options: {
-				gradient: {
-					from: '#D63638',
-					to: '#D63638',
-					fromOpacity: 0.2,
-					toOpacity: 0,
-				},
-				stroke: '#D63638',
-				legendShapeStyle: {
-					color: '#D63638',
-				},
-			},
-		},
 	];
 
+	const formatUnit = ( value: number | string ) => {
+		const num = parseFloat( value as string );
+		if ( [ 'lcp', 'fcp', 'ttfb' ].includes( activeTab ) ) {
+			return +( num / 1000 ).toFixed( 2 );
+		}
+		return num;
+	};
+
+	const displayUnit = () => {
+		if ( [ 'lcp', 'fcp', 'ttfb' ].includes( activeTab ) ) {
+			return __( 's' );
+		}
+		if ( [ 'inp', 'tbt' ].includes( activeTab ) ) {
+			return __( 'ms' );
+		}
+		return '';
+	};
+
+	const isOverall = activeTab === 'overall';
+
 	return (
-		<LineChart
-			data={ _data }
-			withGradientFill
-			height={ 450 }
-			maxWidth={ 1400 }
-			showLegend
-			withLegendGlyph
-		/>
+		<>
+			<VStack direction="row" spacing={ 2 } justify="flex-start">
+				<HStack justify="flex-start">
+					<StatusIndicator speed="good" />
+					<Text size="small">{ __( 'Excellent' ) }</Text>
+					<Text>
+						{ isOverall
+							? sprintf(
+									/* translators: %(to)s is the good threshold */
+									__( '(90–%(to)s)' ),
+									{
+										to: formatUnit( good ),
+									}
+							  )
+							: sprintf(
+									/* translators: %(to)s is the good threshold, %(unit)s is the unit */
+									__( '(0–%(to)s%(unit)s)' ),
+									{
+										to: formatUnit( good ),
+										unit: displayUnit(),
+									}
+							  ) }
+					</Text>
+				</HStack>
+				<HStack justify="flex-start">
+					<StatusIndicator speed="needsImprovement" />
+
+					<Text size="small">{ __( 'Needs Improvement' ) }</Text>
+					<Text>
+						{ isOverall
+							? sprintf(
+									/* translators: %(to)s is the needs improvement threshold */
+									__( '(50–%(to)s)' ),
+									{
+										to: formatUnit( needsImprovement ),
+									}
+							  )
+							: sprintf(
+									/* translators: %(from)s is the good threshold, %(to)s is the needs improvement threshold, %(unit)s is the unit */
+									__( '(%(from)s–%(to)s%(unit)s)' ),
+									{
+										from: formatUnit( good ),
+										to: formatUnit( needsImprovement ),
+										unit: displayUnit(),
+									}
+							  ) }
+					</Text>
+				</HStack>
+				<HStack justify="flex-start">
+					<StatusIndicator speed="bad" />
+
+					<Text size="small">{ __( 'Poor' ) }</Text>
+					<Text>
+						{ isOverall
+							? sprintf(
+									/* translators: %(to)s is the bad threshold */
+									__( '(0-%(to)s)' ),
+									{
+										to: formatUnit( bad ),
+									}
+							  )
+							: sprintf(
+									/* translators: %(from)s is the needs improvement threshold, %(unit)s is the unit */
+									__( '(Over %(from)s%(unit)s)' ),
+									{
+										from: formatUnit( needsImprovement ),
+										unit: displayUnit(),
+									}
+							  ) }
+					</Text>
+				</HStack>
+			</VStack>
+			<LineChart
+				data={ _data }
+				withGradientFill
+				smoothing={ false }
+				maxWidth={ 1400 }
+				renderGlyph={ ( glyphProps ) => {
+					console.log( glyphProps );
+					return <rect width="6" height="6" fill="#3858E9" />;
+				} }
+			/>
+		</>
 	);
 }
